confirms, emails, delete bug fix

+++ app/routes/items.js
@@ -51,13 +51,17 @@ exports.create = function(req, res){
 
 exports.destroy = function(req, res){
   Item.findById(req.params.id, function(item){
-    if(req.session.userId === item.userId.toString()){
-      var cmd = 'rm -rf ' + __dirname + '/../static' + item.photoPath;
-      Item.deleteById(req.params.id, function(){
-        exec(cmd, function(){
-          res.redirect('/users/'+req.session.userId);
+    if(!item.offered){
+      if(req.session.userId === item.userId.toString()){
+        var cmd = 'rm -rf ' + __dirname + '/../static' + item.photoPath;
+        Item.deleteById(req.params.id, function(){
+          exec(cmd, function(){
+            res.redirect('/users/'+req.session.userId);
+          });
         });
-      });
+      }
+    }else{
+      res.redirect('/users/'+req.session.userId);
     }
   });
 };
@@ -101,9 +105,11 @@ exports.accept = function(req, res){
       itemOffer.toggleOffered();
       item.update(function(){
         itemOffer.update(function(){
-          item.sendAcceptEmail();
-          itemOffer.sendAcceptEmail();
-          res.redirect('/users/'+req.session.userId);
+          item.sendAcceptEmail(itemOffer.name, function(){
+            itemOffer.sendAcceptEmail(item.name, function(){
+              res.redirect('/users/'+req.session.userId);
+            });
+          });
         });
       });
     });
